Fix the heuristic for SNES header location detection.
Doesn't directly use code from MAME, because that's C++, but it adapts it into C# so I figure that probably counts it as a derived work and I'd better treat it as such just in case. I like to stay on the good side of open source licenses.

Anyway, this makes the SNES handler actually useful.